Move generation of instance ID into ServerBase
This patch also changes the sequence number to be time-based instead of
constant. This helps with some integration tests where we rely on
restarting a tablet server and assume that the new instance has a higher
sequence number (and thus can re-register).

The timer used isn't an actual wall-clock time, though, so this is
useless in real life. We will need to come back to this relatively
soon, but this unblocks the above-mentioned tests.

Reviewed-on: http://gerrit.ent.cloudera.com:8080/653
Tested-by: jenkins

+++ src/master/ts_descriptor.cc
@@ -61,7 +61,7 @@ MonoDelta TSDescriptor::TimeSinceHeartbeat() const {
   return now.GetDeltaSince(last_heartbeat_);
 }
 
-int32_t TSDescriptor::latest_seqno() const {
+int64_t TSDescriptor::latest_seqno() const {
   boost::lock_guard<simple_spinlock> l(lock_);
   return latest_seqno_;
 }
